chore: move legacy analyzer into ccip pkg
